PreProcessor, IRequestHandler: adjust to new MultiSizeBufferPool
PreProcessor, IRequestHandler: adjust to new MultiSizeBufferPool

Preprocessor now uses a MultiSizeBufferPool with a fixed configuration.
In future tasks, we will allow fetching configuration from a file.
It calls 1st time to getBufferBySubpoolSelector to get
a fixed size buffer. Since pool selector is default, buffer is usually
fetched from the smallest subpool. If requests fail due to insufficient
buffer size, the required size is not returned all the way to
preprocessor which retries and asks the pool for a buffer of that
minimal size. @nd retry should never fail, and pool
should be configured in such a way that a retry is rare enough not to
impact performance.